+++ src/storage-mover/azext_storage_mover/aaz/latest/storage_mover/job_run/_show.py
@@ -22,9 +22,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-07-01",
+        "version": "2025-07-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/projects/{}/jobdefinitions/{}/jobruns/{}", "2024-07-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.storagemover/storagemovers/{}/projects/{}/jobdefinitions/{}/jobruns/{}", "2025-07-01"],
         ]
     }
 
@@ -150,7 +150,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-07-01",
+                    "api-version", "2025-07-01",
                     required=True,
                 ),
             }
@@ -268,7 +268,7 @@ class Show(AAZCommand):
                 serialized_name="itemsUnsupported",
                 flags={"read_only": True},
             )
-            properties.job_definition_properties = AAZObjectType(
+            properties.job_definition_properties = AAZAnyType(
                 serialized_name="jobDefinitionProperties",
                 flags={"read_only": True},
             )
@@ -288,7 +288,7 @@ class Show(AAZCommand):
                 serialized_name="sourceName",
                 flags={"read_only": True},
             )
-            properties.source_properties = AAZObjectType(
+            properties.source_properties = AAZAnyType(
                 serialized_name="sourceProperties",
                 flags={"read_only": True},
             )
@@ -303,7 +303,7 @@ class Show(AAZCommand):
                 serialized_name="targetName",
                 flags={"read_only": True},
             )
-            properties.target_properties = AAZObjectType(
+            properties.target_properties = AAZAnyType(
                 serialized_name="targetProperties",
                 flags={"read_only": True},
             )